central-database/django_requirements.txt
asgiref==3.8.1
Django==5.2.3
django-allauth==64.2.1
dotenv==0.9.9
psycopg2-binary==2.9.10
python-dotenv==1.1.0
qrcode==8.2
Pillow==10.4.0
sqlparse==0.5.3
requests==2.31.0
PyJWT==2.8.0
cryptography>=41.0.0
djangorestframework==3.15.1
gunicorn 
django-cors-headers==4.3.1 